moved PatchInfo logic into utils, so other tools can inspect the system for installed patches
